Socket should to connect to redis specified in .env
Im actually using the redis in docker means another machine it has its own ip address, all are working fine but notifications when deploying are not working, when i debugged in the `socket.js` it is connecting to local redis instance rather than redis that is specified in the .env file. I think it is better to connect to redis which is specified in .env. Generally when using the remote redis it is preferrable to have a password so i have given the option to specify the `redis_password`